requirements.txt
absl-py==2.1.0
altgraph==0.17.4
astunparse==1.6.3
blinker==1.7.0
build==1.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
docutils==0.20.1
filelock==3.14.0
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
huggingface-hub==0.23.1
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==5.1.0
jaraco.functools==4.0.0
jeepney==0.8.0
Jinja2==3.1.3
joblib==1.3.2
keras==3.1.1
keyring==25.1.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
namex==0.0.7
networkx==3.3
nh3==0.2.17
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
pkginfo==1.10.0
protobuf==4.25.3
pycparser==2.22
Pygments==2.18.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pyparsing==3.1.2
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
readme_renderer==43.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.4.1.post1
scipy==1.13.0
SecretStorage==3.3.3
setuptools==70.0.0
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==0.0.1
termcolor==2.4.0
threadpoolctl==3.4.0
timm==1.0.3
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
twine==5.0.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
zipp==3.18.1
